feat(rust): BE-01 Sprint A — 4 deterministic Net Profit Gate components
Components surfaced:
- Component 2 (lp_fees_usd): explicit line-item field in RoiCalculationParams;
  config_aware.rs passes 0.0 in Sprint A (route-leg detail wired in Sprint B).
- Component 3 (price_impact_pct): real V2 impact from reserves when available;
  calc_univ2_price_impact() imported and ready; V3 + reserve-fetch deferred to
  Sprint B. Falls back to max_slippage_pct proxy when price_impact_pct == 0.
- Component 6 (capital_cost_usd): from trading_config.capital_cost_rate_annual_pct
  (default 0.0 APR); flash-loan strategies short-circuit to 0; ETH 12s block time
  constant with per-chain override deferred to Sprint B.
- Component 7 (ops_overhead_usd): from trading_config.ops_overhead_usd_per_attempt
  (default \$0.01); serde default preserves backward compat for existing Redis configs.

Schema: migration 047 adds 2 columns to trading_config with CHECK constraints
  (capital_cost_rate_annual_pct 0-50%; ops_overhead_usd_per_attempt 0-\$100).

Evidence: OpportunityEvidence gains CostBreakdown struct with 7 itemised fields
  + total_cost_usd, populated by ConfigAwareEvaluator and available to dashboard.

ConfigGateOutcome::Evaluated: outcome + evidence boxed (large_enum_variant fix).

Clippy: fixes all pre-existing workspace-wide -D warnings violations across
  sim-ctl, prioritization-spine, recon, relays-client, searcher-rs, shared-rs.

Tests: 9 new unit tests in roi_engine + 55 config_aware + 67 shared-rs = 136 total.
  All pass. calc_net_profit_and_roi remains a pure function; inputs are deterministic.

Per gap analysis 2026-05-08 BE-01. Components 4, 5, 9 in Sprint B/C.

+++ backend/relays-client/src/persistence.rs
@@ -92,7 +92,7 @@ pub async fn persist_execution(pool: &PgPool, r: &ExecutionResult, chain_id: i64
             )
             .bind(relay_name).bind(chain_id)
             .bind(window_start).bind(window_end)
-            .bind(submitted as i32).bind(included as i32).bind(reverted as i32).bind(dropped as i32)
+            .bind(submitted).bind(included).bind(reverted).bind(dropped)
             .execute(&mut *tx)
             .await
             .context("upsert relay_scores")?;
